requirements.txt
aiofiles==0.6.0
blinker==1.4
click==7.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
priority==1.3.0
protobuf==3.15.8
Quart==0.14.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wsproto==1.0.0
